requirements.txt
django
python-telegram-bot==13.15
environs